+++ packages/components/CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Changed
 
+- `Toast`: Alignment reflects RTL (right-to-left) mode
+  ([#3010](https://github.com/porsche-design-system/porsche-design-system/pull/3010))
 - `Carousel`: Pagination can be used for navigation & pagination has more spacing on touch devices
   ([#3003](https://github.com/porsche-design-system/porsche-design-system/pull/3003))
 - `Button`, `Button Pure`, `Switch`, `Checkbox Wrapper`, `Radio Button Wrapper`: optimized announcement of loading state